refactor[ats_ATS240585]: Change item ids to Long

+++ plugins/org.eclipse.osee.ats.rest/src/org/eclipse/osee/ats/rest/internal/config/ReportResource.java
@@ -61,7 +61,7 @@ public class ReportResource {
       Set<Long> modArts = new HashSet<>();
       Set<Long> deletedArts = new HashSet<>();
 
-      Map<Integer, Pair<ChangeItem, Set<ChangeItem>>> artToChanges = new HashMap<>();
+      Map<Long, Pair<ChangeItem, Set<ChangeItem>>> artToChanges = new HashMap<>();
 
       buildArtIdToChangeMap(changes, artToChanges);
       buildLists(artToChanges, newArts, modArts, deletedArts);
@@ -82,7 +82,7 @@ public class ReportResource {
          "application/xml").build();
    }
 
-   private void buildArtIdToChangeMap(List<ChangeItem> changes, Map<Integer, Pair<ChangeItem, Set<ChangeItem>>> artToChanges) {
+   private void buildArtIdToChangeMap(List<ChangeItem> changes, Map<Long, Pair<ChangeItem, Set<ChangeItem>>> artToChanges) {
       for (ChangeItem change : changes) {
          int artId = change.getArtId().getId().intValue();
          ChangeType changeType = change.getChangeType();
@@ -109,9 +109,9 @@ public class ReportResource {
       }
    }
 
-   private void buildLists(Map<Integer, Pair<ChangeItem, Set<ChangeItem>>> artToChanges, Set<Long> newArts, Set<Long> modArts, Set<Long> deletedArts) {
+   private void buildLists(Map<Long, Pair<ChangeItem, Set<ChangeItem>>> artToChanges, Set<Long> newArts, Set<Long> modArts, Set<Long> deletedArts) {
 
-      for (Integer artId : artToChanges.keySet()) {
+      for (Long artId : artToChanges.keySet()) {
          Pair<ChangeItem, Set<ChangeItem>> pair = artToChanges.get(artId);
          ChangeItem artChange = pair.getFirst();
          ModificationType modType = artChange.getNetChange().getModType();
