+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/IdentityManager.java
@@ -23,11 +23,11 @@ public interface IdentityManager {
 
    TransactionId getNextTransactionId();
 
-   int getNextArtifactId();
+   long getNextArtifactId();
 
-   int getNextAttributeId();
+   long getNextAttributeId();
 
-   int getNextRelationId();
+   long getNextRelationId();
 
    GammaId getNextGammaId();
