+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/loader/data/AttributeObjectFactory.java
@@ -25,9 +25,9 @@ import org.eclipse.osee.orcs.core.ds.VersionData;
  */
 public interface AttributeObjectFactory extends VersionObjectFactory {
 
-   <T> AttributeData<T> createAttributeData(VersionData version, Integer id, AttributeTypeGeneric<?> attributeType, ModificationType modType, ArtifactId artId, T value, String uri, ApplicabilityId applicId);
+   <T> AttributeData<T> createAttributeData(VersionData version, Long id, AttributeTypeGeneric<?> attributeType, ModificationType modType, ArtifactId artId, T value, String uri, ApplicabilityId applicId);
 
-   <T> AttributeData<T> createAttributeData(VersionData version, Integer id, AttributeTypeGeneric<?> attributeType, ModificationType modType, ArtifactId artId, ApplicabilityId applicId);
+   <T> AttributeData<T> createAttributeData(VersionData version, Long id, AttributeTypeGeneric<?> attributeType, ModificationType modType, ArtifactId artId, ApplicabilityId applicId);
 
    <T> AttributeData<T> createCopy(AttributeData<T> source);
 }
