+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/util/Address.java
@@ -24,7 +24,7 @@ import org.eclipse.osee.framework.core.enums.TxCurrent;
  */
 public final class Address implements Comparable<Address> {
    private final BranchId branch;
-   private final int itemId;
+   private final Long itemId;
    private final long transactionId;
    private final GammaId gammaId;
    private ModificationType modType;
@@ -34,7 +34,7 @@ public final class Address implements Comparable<Address> {
    private TxCurrent correctedTxCurrent;
    private boolean purge;
 
-   public Address(boolean isBaseline, Long branchId, int itemId, long transactionId, GammaId gammaId, ModificationType modType, ApplicabilityId appId, TxCurrent txCurrent) {
+   public Address(boolean isBaseline, Long branchId, Long itemId, long transactionId, GammaId gammaId, ModificationType modType, ApplicabilityId appId, TxCurrent txCurrent) {
       super();
       this.branch = BranchId.valueOf(branchId);
       this.itemId = itemId;
@@ -107,7 +107,7 @@ public final class Address implements Comparable<Address> {
       return branch;
    }
 
-   public int getItemId() {
+   public Long getItemId() {
       return itemId;
    }
